pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = ""
authors = [
    {name = "Menna Tullah Magdy Taha",email = "menna.tullah.magdy731@gmail.com"}
]
packages = [
    { include = "./tests" },
    { include = "./views" },
    { include = "./background_tasks" },
    { include = "./models" },
    { include = "./services" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "dpytest (>=0.7.0,<0.8.0)",
    "pre-commit (>=4.2.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
